% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{ccor_paired_obj}
\alias{ccor_paired_obj}
\title{Plot the correlation between two variables}
\usage{
ccor_paired_obj(EXdata, obj1, obj2, Group)
}
\arguments{
\item{EXdata}{Data frame containing the variables and grouping column.}

\item{obj1, obj2}{Character column names of the variables to compare.}

\item{Group}{Character name of the column used to colour points.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Computes a Pearson correlation and linear regression between two columns and
colours samples by a grouping column.
}
