% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{plot_roc_curve}
\alias{plot_roc_curve}
\title{Plot time-dependent survival ROC curves}
\usage{
plot_roc_curve(risk.df)
}
\arguments{
\item{risk.df}{Data frame containing survival \code{Time}, binary \code{Status}, and
\code{Risk_score} columns.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object containing the three ROC curves.
}
\description{
Calculates survival ROC curves at one, three, and five years and displays
their AUC values in a combined plot.
}
